requirements.txt
playwright==1.40.0
pytest==8.0.0
pytest-playwright==0.4.3
pytest-html==4.1.0
pytest-xdist==3.5.0
pytest-rerunfailures==14.0
pytest-timeout==2.2.0
pytest-env==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1
